Small compatibility changes
These are changes made to the models created by Sandra. They include required fields, default values and a reformatting of the  collection to reflect the layout with which the new JSON data has been printed. **NOTE**: Read/Write order in JSON does not matter, so this was purely aesthetic for checking the model vs. the printed data